+++ src/hackrf-sweep/src/test/java/shared/mvc/ModelValueTest.java
@@ -3,6 +3,7 @@ package shared.mvc;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Consumer;
 
 import org.junit.jupiter.api.Test;
 
@@ -15,10 +16,10 @@ class ModelValueTest {
     void testBasicSetGetAndListener() {
         ModelValueInt mv = new ModelValueInt("test", 10);
         assertEquals(10, mv.getValue());
-        assertEquals("test", mv.getName());
+        assertEquals("test", mv.toString());
 
         AtomicInteger calls = new AtomicInteger(0);
-        mv.addListener(v -> calls.incrementAndGet());
+        mv.addListener((Consumer<Integer>) v -> calls.incrementAndGet());
 
         mv.setValue(20);
         assertEquals(20, mv.getValue());
@@ -41,26 +42,25 @@ class ModelValueTest {
         assertEquals(1, i.getStep());
         assertEquals(0, i.getMin());
         assertEquals(10, i.getMax());
+        assertThrows(IllegalStateException.class, () -> i.setValue(11));
+        assertEquals(5, i.getValue());
     }
 
     @Test
-    void testListenerRemovalAndMultipleListeners() {
+    void testMultipleListenersAndEqualsNoOp() {
         ModelValueInt mv = new ModelValueInt("val", 0);
-        java.util.concurrent.atomic.AtomicInteger c1 = new java.util.concurrent.atomic.AtomicInteger(0);
-        java.util.concurrent.atomic.AtomicInteger c2 = new java.util.concurrent.atomic.AtomicInteger(0);
+        AtomicInteger c1 = new AtomicInteger(0);
+        AtomicInteger c2 = new AtomicInteger(0);
 
-        ModelValue.Listener<Integer> l1 = v -> c1.incrementAndGet();
-        ModelValue.Listener<Integer> l2 = v -> c2.incrementAndGet();
+        mv.addListener((Consumer<Integer>) v -> c1.incrementAndGet());
+        mv.addListener((Runnable) c2::incrementAndGet);
 
-        mv.addListener(l1);
-        mv.addListener(l2);
         mv.setValue(1);
         assertEquals(1, c1.get());
         assertEquals(1, c2.get());
 
-        mv.removeListener(l1);
-        mv.setValue(2);
+        mv.setValue(1);
         assertEquals(1, c1.get());
-        assertEquals(2, c2.get());
+        assertEquals(1, c2.get());
     }
 }